Translating ἰδού as "behold" rather than "wow" etc
"Behold" is a more literal rendering of ἰδού, has the familiarity of agreement with most other translations, and is less informal than "wow."  While "wow" might carry some of the connotation of the word, it doesn't have the right denotation. Furthermore, Pickering is not always consistent in his translation of this word, sometimes rendering it "but," "then," "take note." While "take note" is a fairly decent translation, "Behold" has a ring of familiarity with other translations and gives a more literal rendering. Therefore I have changed every occurrence of ἰδού with "behold."
Josh, could you do a global replace of all occurrences of "wow" with "behold." Otherwise I will manually change 1:7,18; 3:11,20; 4:1,2; 5:5,6; 6:12 [I already changed 6:2,5,8], 7:9; 9:12; 11:14; 12:3; 14:1