Adding "and" to 14:6
This is because "earth" refers to Israel, and there is a και immediately after it. Pickering no doubt interpreted the και  as an "even" or "that is" (no doubt because he fails to see that γῆς is a reference to Israel. However, I take it as a sequence of και words indicating that this message went beyond Israel.